backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
pymongo
psycopg2-binary



#npm install bootstrap
